dhcp-proxy: do not remove socket when running under systemd
Because we use socket activation systemd manages the socket for us, if
we remove the file after the timeout all new request will fails as
systemd will only recreate the socket if you restart the socket unit.

+++ src/commands/dhcp_proxy.rs
@@ -240,6 +240,8 @@ pub async fn serve(opts: Opts) -> NetavarkResult<()> {
     // Watch for signals after the uds path has been created, so that the socket can be closed.
     handle_signal(uds_path.clone()).await;
 
+    let mut is_systemd_activated = false;
+
     // check if the UDS is a systemd socket activated service.  if it is,
     // then systemd hands this over to us on FD 3.
     let uds: UnixListener = match env::var("LISTEN_FDS") {
@@ -248,6 +250,7 @@ pub async fn serve(opts: Opts) -> NetavarkResult<()> {
                 error!("Received more than one FD from systemd");
                 return Ok(());
             }
+            is_systemd_activated = true;
             let systemd_socket = unsafe { stdUnixListener::from_raw_fd(3) };
             systemd_socket.set_nonblocking(true)?;
             UnixListener::from_std(systemd_socket)?
@@ -300,7 +303,11 @@ pub async fn serve(opts: Opts) -> NetavarkResult<()> {
         _ = &mut server => {},
     };
 
-    fs::remove_file(uds_path);
+    // Make sure to only remove the socket path when we do not run socket activated,
+    // otherwise we delete the socket systemd is using which causes all new connections to fail.
+    if !is_systemd_activated {
+        fs::remove_file(uds_path);
+    }
     Ok(())
 }
